modified:   lssn-9-multer-books/controllers/auth/setAvatar.js 	modified:   lssn-9-multer-books/helpers/index.js 	new file:   lssn-9-multer-books/helpers/manipulate.js
