+++ lssn-9-multer-books/controllers/auth/setAvatar.js
@@ -1,23 +1,10 @@
-const patch = require("path");
-const Jimp = require("jimp");
+const path = require("path");
 const fs = require("fs/promises");
 
 const { basedir } = global;
 const { User } = require(`${basedir}/models/users`);
-const avatarsDir = patch.join(basedir, "public", "avatars");
-
-const manipulate = (tempPath) => {
-  Jimp.read(tempPath)
-    .then((img) => {
-      return img
-        .resize(250, 250) // resize
-        .quality(69) // set JPEG quality
-        .write("temp/pict.jpg"); // save
-    })
-    .catch((err) => {
-      console.error(err);
-    });
-};
+const { manipulate } = require(`${basedir}/helpers`);
+const avatarsDir = path.join(basedir, "public", "avatars");
 
 const setAvatar = async (req, res) => {
   try {
@@ -26,10 +13,10 @@ const setAvatar = async (req, res) => {
     manipulate(tempPath);
     const [extension] = originalname.split(".").reverse();
     const newName = `${_id}.${extension}`;
-    const uploadPath = patch.join(avatarsDir, newName);
+    const uploadPath = path.join(avatarsDir, newName);
     await fs.rename("temp/pict.jpg", uploadPath);
     await fs.unlink(req.file.path);
-    const avatarURL = patch.join("avatars", newName);
+    const avatarURL = path.join("avatars", newName);
     await User.findByIdAndUpdate(_id, { avatarURL });
     res.json({ avatarURL });
   } catch (error) {
