+++ lssn-9-multer-books/helpers/index.js
@@ -1,4 +1,5 @@
 const createError = require("./createError");
 const ctrlWrapper = require("./ctrlWrapper");
+const manipulate = require("./manipulate");
 
-module.exports = { createError, ctrlWrapper };
+module.exports = { createError, ctrlWrapper, manipulate };
